Добавила картинки, изменила значение рейтинга на дабл

+++ server/models/models.js
@@ -20,7 +20,7 @@ const  Device = sequelize.define('device',{
     id: {type:DataTypes.INTEGER,primaryKey:true, autoIncrement:true},
     name:{type:DataTypes.STRING,unique: true, allowNull:false},
     price:{type:DataTypes.INTEGER,allowNull:false},
-    rating:{type:DataTypes.INTEGER,defaultValue:0},
+    rating:{type:DataTypes.DOUBLE,defaultValue:0},
     img:{type:DataTypes.STRING,allowNull:false},
     id_api: {type:DataTypes.STRING,unique: true, allowNull:false},
 })
